ARM: OMAP: remoteproc: add reset handling
There are recent changes in the kernel code that took the
reset control portion for devices with physical reset lines
out of the omap_hwmod/omap_device code. This requires that
the drivers handle the reset lines themselves. The remoteproc
driver belong to this class of drivers.

This additional logic is added into the platform portion of
the remoteproc enable/shutdown functions, together with the
standard omap_device_enable/shutdown functions. This will
allow the design to scale well to DSP and IPU, which may
have different number of reset lines. There is no need to
proliferate separate pdata functions just for controlling
the resets.

The current code also releases the core1 reset to have the
IPU functional in SMP-mode, but this can be fixed later if
a need arises to run IPU in non-SMP mode.

Signed-off-by: Suman Anna <s-anna@ti.com>