remoteproc/omap: request a timer(s) for remoteproc usage
The remote processors in OMAP are equipped with internal timers,
like the internal SysTick timer in a Cortex M3/M4 NVIC or the
CTM timer within Unicache in IPU & DSP. However, these timers
are gated when the processor subsystem clock is gated, making
them rather difficult to use as OS tick sources. They will not
be able to wakeup the processor from any processor-sleep induced
clock-gating states.

This can be avoided by using an external timer as the tick source,
which can be controlled independently by the core remoteproc
driver code, but still allowing the processor subsystem clock to
be auto-gated when the remoteproc cores are idle.

This patch adds the support for remote processors to request a
timer(s) to be used by the remoteproc. The timers are enabled and
disabled in line with the enabling/disabling of the remoteproc.
The timer data is not mandatory if the advanced device management
features are not required.

The current design relies on OMAP dmtimer api, which are only
present in machine-specific layers, but this needs to be
revisited as and when the dmtimer code moves under the /drivers
folder.

The following timers are used as the tick sources for OMAP4+
remoteprocs, these are picked through the platform data.
	DSP : GPT5
	IPU : GPT3

Additional timer(s) can be used in similar fashion for purposes
like watchdog functionality or as a tick source to the second
processor in IPU (in non-SMP mode).

NOTE: If the gptimer is already in use by the time IPU and/or
DSP are loaded, the processors will fail to boot.

Signed-off-by: Suman Anna <s-anna@ti.com>