ARM: OMAP4: hwmod_data: add mmu interrupts back on 3.11
The OMAP4 hwmod file has been heavily cleaned up in 3.11-rc1.
This removed the IRQ information for most of the hwmods, and
is represented through DT data. The OMAP iommu driver is not
yet converted to DT, and hence the driver probe fails.

Add this data back to be able to test remoteproc code.

Signed-off-by: Suman Anna <s-anna@ti.com>