remoteproc/omap: remove the oh_name_opt field from platform data
The oh_name_opt field was primarily defined to identify the hwmod for
the second cpu in a dual Cortex-M3/M4 IPU processor sub-system. This
hwmod entry is no longer defined usually, but rather a single hwmod
with two reset lines is used to represent both the processors in the
IPU sub-system. A single firmware image (either in SMP-mode or a combined
image for non-SMP mode) is used, with both the resets released together
always as part of the device management. Any power management and
recovery aspects require that both the processors be managed as one
entity due to the presence of shared MMU and unicache within the IPU
sub-system.

Signed-off-by: Suman Anna <s-anna@ti.com>