remoteproc: add 64-bit ELF loader support code
The current remoteproc core has supported only 32-bit remote
processors and provides a default 32-bit ELF loader code for
parsing and loading the corresponding firmwares, irrespective
of its native 32-bit or 64-bit Linux OS. Add a new remoteproc
core file that provides equivalent loader functionality to
support 64-bit remote processors using 64-bit ELF firmwares.

The current file is created by using the 32-bit ELF loader
code as a foundation and using the appropriate 64-bit structures.
This is done in a separate file to not disturb the current
32-bit loader code, and without introducing additional changes
into the remoteproc core (Both 32-bit and 64-bit firmwares needs
to be supported irrespective of the 32-bit/64-bit Linux OS, which
requires some runtime flag detection rather than a pre-compiler
detection). The remoteproc platform drivers are expected to
overwrite the required rproc_ops with the 64-bit versions.

NOTE:
The rproc_elf64_get_boot_addr() currently returns a 32-bit entry
point assuming all the higher-bit addresses as zero, to reuse the
rproc ops. This is a temporary solution and needs to be fixed
up in the future.

Signed-off-by: Suman Anna <s-anna@ti.com>